feat: Add support for random subdomain in HTTP tunnel
This commit adds support for a random subdomain in the HTTP tunnel feature. The `Payload` enum in the `event.rs` file now includes a `random_subdomain` field. When this field is set to true, the server will assign a random subdomain for the HTTP tunnel. This provides more flexibility and options for configuring tunnels.

Note: This commit message is based on the analysis of the code changes and recent repository commits.