feat: Add support for TCP and UDP tunnels with random subdomain and remote port (#43)
This commit adds support for TCP and UDP tunnels with the ability to assign a random subdomain and remote port. The `Tunnel` struct now includes methods for creating TCP and UDP tunnels with configurable options such as random subdomain and remote port. This provides more flexibility and options for configuring tunnels.

Note: This commit message is based on the analysis of the code changes and recent repository commits.